chore: Add Utility Makefile
Right now, this is just a wrapper around the utility script that runs
Terraform.

I also refactored the up script so that it could accept a version
number, without having to download the binary manually from S3.

The Makefile is just the way I'm used to using Terraform, and to make
things a bit more convenient.  Use of it is completely optional.