fix: Pass `--local-addr` to `sn_node`
The `sn_node` CLI now defaults to `127.0.0.1` for local address, rather
than `0.0.0.0`, so we need to supply a value if we want external
reachability. We use the node's actual IP rather than `0.0.0.0` due to a
curious behaviour in `sn_node` where it may use the local address as the
public address.