feat(gha-support): add working dir variable
when running the tool on GH actions all files are not created in the
current directory. They are created at $GITHUB_ACTION_PATH. This commit
adds a working_dir variable to the terraform scripts which uses the
GITHUB_ACTION_PATH variable to create and access the files if specified
and defaults to the current dir if undefined (eg. while running locally)