refactor: use single scripts to init nodes
Rather than several scripts to setup and run the node, which were distributed across files and
inline provisioner fields (and the inline scripts are not the best for readability), one
`init-node.sh` script is now used for the main node, and `init-client-node.sh` for the client test
runner.

The first thing to say is, the scripts solved a problem I was having with the apt package manager.
I've documented the problem and solution with comments in the script, so I won't repeat those here,
and just say it solves random provisioning failures by using a retry loop.

The other important aspect is, the main node script now requires a URL for the node, which will
point at an archive in an S3 bucket which contains the node. This will either be the 'latest' node,
a specific version that was released, or a custom built binary from the local machine, which was
uploaded prior to the Terraform run. The `up.sh` script was modified to accommodate either of those
scenarios.

The URL change removed the need for the Terraform `node_bin` variable, which was being used in
an awkward manner with the provisioners.

Other misc changes:
* Introduce the ability to toggle the new client node type by using a count. Default is 0.
* Update the README for new `up.sh` script arguments.
* Remove the `working_dir` variable, which didn't appear to serve any purpose and made scripts more
  verbose.
* Use consistent indent size of 2 for Terraform manifests.