feat: suppress output from apt
The installation of various packages produces an enormous amount of text output that is largely of
no interest, and it makes the nightly logs difficult to sift through. Note that I tried various
switches like `-qq`, but there was nothing that worked across different versions of Ubuntu, so I
ended up just redirecting the output to `/dev/null`.

You'd only really be interested in seeing the output when the installs fail, but I'd say in that
case you can turn the output back on to see what is going on. If the package is on apt it should be
pretty stable anyway and an unlikely source of errors.

There are also a couple of changes to the Makefile targets here just to make sure all information
relating to the previous testnet are cleared out.