feat(node): provide genesis key from genesis node
When the genesis node starts up, the genesis key gets output to file by finding it from the logs.
After the Terraform run it will then be copied from the genesis node to the local machine and also
uploaded to S3, as per the IP lists and the prefix map.

Having the genesis key can be useful for basic debugging. We had seen some errors that potentially
indicated a mismatch between the genesis key on the node and the one being used by the client test
runner process.

The down script removed the prefix map and genesis information from S3, but I've taken this out,
since having access to this information could be useful, particularly after a failed nightly run.

There was also a local-exec provisioner that ran to download these files from S3 if they didn't
exist on the local machine. I couldn't see what the purpose of this was, so I removed it. That
information will get generated by the `up.sh` script.