feat(debug): script for pulling logs
Based on the `logs` script, this one is very similar, it just operates in a synchronous way rather
than launching rsync in the background.

It's intended to be used mainly for debugging the nightly, so we will want to wait till all the logs
have been downloaded before moving on to the next step.