fix(script): continue on error for log sync
There was a problem with a nightly run where rsync returned an error due to a 'vanished' file, which
is one that was on its transfer list initially, but was subsequently deleted before it actually came
to be transferred. Even though rsync treats this as a warning, it still returns an non-zero exit
code, and you can't change that behaviour using an argument. This caused the run to fail before the
rest of the logs were transferred or the testnet was killed.

I tried a few different methods, but the simplest solution may just be to let the script continue on
errors, so `set -e` is removed.