fix: retain use of node bin path argument
The NODE_BIN_PATH argument had been changed to assume a custom binary located at
"workspace/${testnet_channel}/sn_node". However, we still need to be able to use the custom location
for use with the testnet custom action. I also personally use full paths to node binaries located in
the directory where I have `safe_network` code checked out.

This change will check if the `NODE_BIN_PATH` argument is set, and if not, it will look for a binary
in the workspace directory.