feat: run a testnet on aws
Use EC2 instances to run a testnet on AWS.

The responsibility of this repo is just to deal with creating the testnet. We assume that existing
VPC infrastructure has already been created elsewhere and we supply subnet and security group IDs
from that infrastructure.

Terraform is used to create EC2 instances for running nodes and Ansible is used to provision them
with the node setup. This means we can keep the Terraform setup extremely simple and make use of
Ansible's features like Jinja templating to setup the machines.

Rather than using scripts directly, Just (a modern alternative to Make) is used for creating and
working with the testnet.

At the moment this setup doesn't support supplying a custom node binary and the nodes communicate
with each other using IP addresses that are local to the VPC.