feat: create a client machine for running tests
Use the `enable_client` variable to control the creation of a client machine from which you can run
a set of tests against the testnet. This machine uses the same AMI and type as the one we use for
our self hosted CI jobs.

The Justfile targets were refactored to smaller functions to reduce duplication.