feat: support digital ocean with new mechanism
The new mechanism for creating testnets with AWS is now extended to support Digital Ocean as a
provider.

A simple Terraform setup is provided that just creates a genesis node and all the other nodes. As
with the AWS setup, Ansible is then used to provision the nodes. The setup between the two is mostly
the same, but with two notable differences:

* On Digital Ocean, no additional disks are supplied for logging and data, because you get a fairly
  big 60GB allocation for the main disk.
* The node setup is a bit simpler on Digital Ocean because its network interface has the public IP
  address.

With the introduction of Ansible, the prerequisite tool setup was getting out of hand, so a
container has been provided which packages all the tools necessary to perform the runs. So now, the
only dependency required on the host machine is Docker.