chore: move aws directory to top level
The AWS directory is moved to the top level and the previous files at this level are all cleared
out.

We don't need the old Terraform definitions any more, and a lot of the scripts just didn't apply any
more.

This way we can start fresh and add new scripts (or playbooks) as and when we need them.