feat: use `up` and `down` wrappers
The definition of the container was changed such that arbitrary commands can be provided to run in
the container's Bash shell. This removes the requirement to use the shell inside the container for
running `just` commands. It seemed appropriate to rename the `runner.sh` script to `run.sh`, since
it now runs arbitrary commands.

In addition to `run.sh`, `up.sh` and `down.sh` scripts were provided, which can be used as simple
wrappers for bringing up or tearing down a testnet without using any `just` commands directly.

The `run.sh` script should still be used for doing things like retrieving logs.

The container definition was also modified to revert back to using the single `RUN` instruction, as
per best practice.