feat: support custom local safenode binary
A locally built binary can be used by supplying an extra argument to the `up.sh` script.

It will be copied to the `bin` directory in this repository, which means it will be available inside
the container and the `just` targets can check to see if it exists, and if it does, use that binary
instead of pulling the latest one from S3.