feat: use platform specific dockerfiles
It was observed that the container build performance was very poor on macOS because of a CPU
mistmatch in the base container. Therefore, we use two different container definitions.

A utility script is provided to detect the OS and perform the build based on that.

Also changed the Rust install mechanism to pass the arguments directly rather than using an
intermediately saved binary.