chore: support creating testnets non-interactively
Make some accommodations for running the testnet creation process in a Github Actions workflow.

The `--interactive` argument is removed from the `run` commands because this seems to be problematic
when attempting to run the process in a non-interactive environment like on Github Actions.

These arguments were initially added when the container was intended to be used in interactive mode,
and now don't seem necessary when we are passing commands to be executed.

The use of the `--tty` argument is retained for the second `run` command because Ansible still seems
to require the use of it.

We now also make a fully qualified reference to the container using
`jacderida/sn_testnet_tool:latest`. The container has been uploaded to Dockerhub, which prevents
the workflow from having to perform a `docker build`, which could potentially take a long time on
the GHA infrastructure.