feat: run multiple node services on single vm
Using systemd it is actually quite simple to run multiple instances of the same service.

It has also been made simple by recent changes we've made in the node for providing default
directories for root and logging, and the fact that the node can select an unused port for us.