feat: provide `resource-usage` script
The script is copied from a forked branch, with some minor modifications.

We can run it through the `ansible` command, which provides the inventory and runs the commands in
parallel.

The `safe` user on Digital Ocean was also modified so that we can grant SSH access to this user,
through the same set of keys provisioned for the `root` user. It's easier if the script executes as
the `safe` user because the services are running as that user.