feat: provide `resource_usage` role
The resource usage script runs on each machine in the testnet and outputs its logs to
`~/.local/share/safe/node/resource-usage.log`. It is running as the `safe` user.

A little modification was made to the script to get it to execute in an infinite loop.

The `resource-usage` target in the Justfile was removed because it wouldn't really work if the
script is running in an infinite loop.