feat: provide log forwarding role
Uses Logstash to forward the logs from the nodes and a resource usage script which has been setup to
run as a service. They are forwarded to an S3 bucket. Due to the way that Logstash works, each log
is forwarded in parts, with all the parts stored in a directory in the bucket. The rotation period
is time based, occurring once every 10 minutes.

Some modifications were made to the `safe` user so the Logstash service could access the log files
for uploading them.