chore: provide target to clear previous logs
If a testnet with the same name has ran previously, the logs for that testnet may still exist. If
this is the case, we refuse to proceed until the user has cleared those logs out. This gives the
user the chance to retrieve them before they are cleared, rather than automatically removing them.

Once the logs have been retrieved, or if they are not required, the user can run the `clean-logs`
target to delete them. The new testnet run can now proceed.

The log output location in the S3 bucket was changed slightly just to accommodate this.

The `logs.yml` playbook has been removed, since this retrieval mechanism no longer applies or is
really necessary now that we've got the logs going to S3.